+++ spec/attachie/s3_driver_spec.rb
@@ -55,6 +55,16 @@ RSpec.describe Attachie::S3Driver do
         tempfile.close(true)
       end
     end
+
+    it "raises an Attachie::ItemNotFound when the object does not exist" do
+      tempfile = Tempfile.new
+
+      begin
+        expect { driver.download("unknown", "bucket", tempfile.path) }.to raise_error(Attachie::ItemNotFound)
+      ensure
+        tempfile.close(true)
+      end
+    end
   end
 
   describe "#store_multipart" do
@@ -84,6 +94,10 @@ RSpec.describe Attachie::S3Driver do
       ensure
         driver.delete("name", "bucket")
       end
+    end
+
+    it "returns true even when the object does not exist" do
+      expect(driver.delete("unknown", "bucket")).to eq(true)
     end
   end
 
@@ -130,6 +144,10 @@ RSpec.describe Attachie::S3Driver do
       ensure
         driver.delete("name.txt", "bucket")
       end
+    end
+
+    it "raises an Attachie::ItemNotFound when the object does not exist" do
+      expect { driver.info("unknown", "bucket") }.to raise_error(Attachie::ItemNotFound)
     end
   end
 end